social_media/requirements.txt
asgiref==3.6.0
Django==4.1.7
djangorestframework==3.14.0
mysqlclient==2.1.1
django-mysql==3.10.0
Pillow==9.5.0
python-decouple==3.8
pytz==2023.3
sqlparse==0.4.3
tzdata==2023.3
django-cors-headers==3.10.1
